修改

+++ NetFrame/Tool/TimeEventModel.cs
@@ -7,14 +7,14 @@ namespace NetFrame.Tool
     public class TimeEventModel
     {
         /// <summary>
-        /// 执行或时间（单位 秒）
+        /// 执行或时间
         /// </summary>
         public long Excute_time;
 
         /// <summary>
-        /// 时间间隔（单位 秒）
+        /// 时间间隔
         /// </summary>
-        public int Wait_time;
+        public long Wait_time;
 
         public int count;
 
@@ -28,7 +28,7 @@ namespace NetFrame.Tool
         /// <param name="time">执行间隔 单位 秒</param>
         /// <param name="count">执行次数 -1 表示无限循环 此时需要外部移除任务</param>
         /// <param name="de">触发事件</param>
-        public TimeEventModel(int time, int count, Action de) {
+        public TimeEventModel(long time, int count, Action de) {
 
             this.Wait_time = time * 10000 * 1000;//把秒数转换成 ticks 的单位长度
             this.Excute_time = DateTime.Now.Ticks+Wait_time;
@@ -44,7 +44,7 @@ namespace NetFrame.Tool
         /// <param name="de">触发事件</param>
         public TimeEventModel(float time, int count, Action de) {
 
-            this.Wait_time = (int)(time * 10000 * 1000);//把秒数转换成 ticks 的单位长度
+            this.Wait_time = (long)(time * 10000 * 1000);//把秒数转换成 ticks 的单位长度
             this.Excute_time = DateTime.Now.Ticks + Wait_time;
             this.count = count;
             this.de = de;
